requirements.txt
Flask==0.9
Jinja2==2.6
WTForms==1.0.2
Werkzeug==0.8.3
distribute==0.6.27
facebook-sdk==0.4.0
flask-peewee==0.6.1
gevent==0.13.8
greenlet==0.4.0
peewee==2.0.4
python-mpd==0.3.0
requests==0.14.2
simplejson==2.6.2
wsgiref==0.1.2
wtf-peewee==0.2.0
